Correctifs: token + injections SQL local storage
- Changement de fonctionnement. Quand un compte est créé ou qu'un utilisateur se déconnecte, on généré aléatoirement un token qu'on stocke dans la db. Cela permet d'éviter qu'entre une déco et une reco, que quelqu'un puisse se connecter avec le token de la session d'avant

- Quand on regarde ce qu'il y a dans le local storage, on regarde qu'il n'y a pas d'injections SQL dedans vu qu'on fait après un appel à la DB (lors du middleware)

+++ ephecReservation/client/src/pages/SignOut/SignOut.tsx
@@ -1,5 +1,7 @@
 import { IonContent, IonHeader, IonPage, IonTitle, IonToolbar, IonButton } from '@ionic/react';
+//importation des autres fichiers
 import './SignOut.css';
+import config from "../../config.json";
 //hook pour check si il y a des données
 import useAuthentication from "../../hooks/checkAuthentication";
 
@@ -8,7 +10,40 @@ const SignOut: React.FC = () => {
   useAuthentication();
 
   function signOut() {
-    localStorage.clear(); // clear all data in local storage
+    let responseBox = document.getElementById('sign_out_response_box')
+    //on change le token avec un généré aléatoirement
+    fetch(config.API_URL + "/auth/token", {
+      method: 'PUT',
+      headers: {
+        'Content-type': 'application/json',
+      },
+      body: (
+        JSON.stringify({
+          upn: localStorage.getItem('upn')
+        }
+        )
+      ),
+    }).then(function (res) {
+      if (res.status === 200) {
+        localStorage.clear(); // clear all data in local storage
+        if (responseBox !== undefined && responseBox !== null) {
+          //si la deuxième requête se passe bien
+        responseBox.innerHTML = "<p id='success_response'>Vous êtes déconnecté.</p>";
+        }
+      }
+      else{
+        //si la deuxième requête se passe pas bien
+        if (responseBox !== undefined && responseBox !== null) {
+          responseBox.innerHTML = "<p id='failed_response'>Un problème est survenu.<br/>Veuillez réessayez plus tard.</p>";
+          }
+      }
+
+    }).catch(function (res) {
+      //si la deuxième requête se passe pas bien
+      if (responseBox !== undefined && responseBox !== null) {
+        responseBox.innerHTML = "<p id='failed_response'>Un problème est survenu.<br/>Veuillez réessayez plus tard.</p>";
+      }
+    })
 
   }
 
@@ -23,6 +58,8 @@ const SignOut: React.FC = () => {
         <div id="sign_out">
           <p>Vous nous quittez déjà ?</p>
           <IonButton color='danger' onClick={() => signOut()}>Déconnexion</IonButton>
+          <div id="sign_out_response_box">
+          </div>
         </div>
 
       </IonContent>
